feat(db): fix float instead of uint8

+++ ee/backend/pkg/db/clickhouse/connector.go
@@ -397,8 +397,8 @@ func (c *connectorImpl) InsertWebClickEvent(session *sessions.Session, msg *mess
 	if msg.Label == "" {
 		return nil
 	}
-	var nX *uint8 = nil
-	var nY *uint8 = nil
+	var nX *float32 = nil
+	var nY *float32 = nil
 	if msg.NormalizedX != 101 && msg.NormalizedY != 101 {
 		// To support previous versions of tracker
 		if msg.NormalizedX <= 100 && msg.NormalizedY <= 100 {
@@ -407,9 +407,9 @@ func (c *connectorImpl) InsertWebClickEvent(session *sessions.Session, msg *mess
 		}
 		normalizedX := float32(msg.NormalizedX) / 100.0
 		normalizedY := float32(msg.NormalizedY) / 100.0
-		nXVal := uint8(normalizedX)
+		nXVal := normalizedX
 		nX = &nXVal
-		nYVal := uint8(normalizedY)
+		nYVal := normalizedY
 		nY = &nYVal
 	}
 	if err := c.batches["clicks"].Append(
